webapp - refactor: ProfitLossAsset 컴포넌트 관련 스타일을 Dashboard/styles.ts 에서 Dashboard/ProfitLossAsset/styles.ts로 옮김
