+++ packages/webapp/src/pages/Dashboard/styles.ts
@@ -1,13 +1,11 @@
 import styled from 'styled-components';
 import Card from '@components/Card';
-import { priceColorMixin } from '@styles/Mixins';
 
 interface ItemIconBgColorProp {
 	bgColor: 'gray' | 'blue';
 }
 
 export const ITEM_UPPER_LOWER_PADDING_PX = 8;
-const PROFIT_LOSS_ASSET_ITEM_HEIGHT_PX = 134;
 const PIE_CHART_CONTAINER_HEIGHT_PX = 400;
 export const CANVAS_PADDING_PX = 10;
 
@@ -53,40 +51,6 @@ export const PortfolioSelectContainer = styled.div`
 	& > select {
 		margin: 0;
 	}
-`;
-
-export const ProfitLossAssetContainer = styled.section`
-	display: flex;
-	justify-content: space-evenly;
-	gap: 30px;
-`;
-
-export const ProfitLossAssetItem = styled(Card)`
-	padding: ${ITEM_UPPER_LOWER_PADDING_PX}px 0;
-	position: relative;
-	display: flex;
-	flex-direction: column;
-	height: ${PROFIT_LOSS_ASSET_ITEM_HEIGHT_PX}px;
-`;
-
-export const ProfitLossAssetAmount = styled.div`
-	${priceColorMixin};
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	margin-top: 17px;
-	padding: 0 5%;
-	font-size: 24px;
-	font-weight: 500;
-`;
-
-export const ProfitLossAssetPercent = styled.div`
-	${priceColorMixin};
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	text-align: center;
-	margin-top: 14px;
 `;
 
 export const DividendPieChartContainer = styled(Card)`
